Scrubby update 3
Slowly dialing in on basic creation of deck and adding one flashcard at a time through "deck management"(will eventually change that to allow for multiple additions, moving, edits, etc.

Currently working on "studying" with the in tent of selecting a deck and then having one face showing with a flip and next option.  Getting weird behavior with a deck title of "D10" always routing incorrectly.

Also just watched another tutorial while I was on the treadmill.  T his guy was writing a basic Todo list using a databvase similar to techwithtims tutorial that wrote a notes page and where I started.  Techwithtim used .js with button onclicks to interface with the database... while the todo list tutorial never used .js and did it directly to Python with hrefs.  Guy on Reddit pointed out I shouldn't be doing onclicks as well so got to figure that out.